Stor del av indexsidan klar

+++ NineMuses/Models/VideoModel.cs
@@ -7,7 +7,7 @@ namespace NineMuses.Models
 {
     public class VideoModel
     {
-        public int UserID { get; set; }
+        public long UserID { get; set; }
         public UserModel User { get; set; }
         public int VideoID { get; set; }
         public string Thumbnail { get; set; }
@@ -17,13 +17,13 @@ namespace NineMuses.Models
         public int Views { get; set; }
         public DateTime UploadDate { get; set; }
 
-        public string GetUserName 
-        { 
+        public string GetUserName
+        {
             get
             {
                 return this.User != null ? this.User.Username : null;
             }
-            
+
         }
     }
 }
